[api] [ranks] 종합순위 api 추가
