+++ opgc/api/githubs/urls.py
@@ -1,7 +1,6 @@
 from django.conf.urls import url
 
-from api.githubs.views import GithubUserViewSet, OrganizationViewSet, RepositoryViewSet, LanguageViewSet, \
-    TierRankViewSet, UserRankViewSet
+from api.githubs.views import GithubUserViewSet, OrganizationViewSet, RepositoryViewSet, LanguageViewSet
 
 app_name = 'githubs'
 
@@ -24,20 +23,9 @@ language_list = LanguageViewSet.as_view({
     'get': 'list',
 })
 
-tier_list = TierRankViewSet.as_view({
-    'get': 'list',
-})
-
-user_rank_list = UserRankViewSet.as_view({
-    'get': 'list',
-})
-
-
 urlpatterns = [
     url(r'^users/(?P<username>[-\w]+)/$', github_user_list, name='github_user_list'),
     url(r'^users/(?P<user_pk>\d+)/organizations/$', organization_list, name='organization_list'),
     url(r'^users/(?P<user_pk>\d+)/repositories/$', repository_list, name='repository_list'),
     url(r'^languages/$', language_list, name='language_list'),
-    url(r'^tier/$', tier_list, name='tier_list'),
-    url(r'^user_rank/$', user_rank_list, name='user_rank_list'),
 ]
