+++ opgc/api/githubs/views.py
@@ -1,13 +1,12 @@
 from datetime import timedelta, datetime
 
-from django_filters.rest_framework import DjangoFilterBackend
 from rest_framework import viewsets, mixins, exceptions
 from rest_framework.response import Response
 
 from api.exceptions import NotExistsGithubUser, RateLimitGithubAPI
 from api.githubs.serializers import GithubUserSerializer, OrganizationSerializer, RepositorySerializer, \
-    LanguageSerializer, TierSerializer
-from api.paginations import IdOrderingPagination, TierOrderingPagination, UserRankOrderingPagination
+    LanguageSerializer
+from api.paginations import IdOrderingPagination
 from apps.githubs.models import GithubUser, Organization, Repository, Language
 from utils.exceptions import GitHubUserDoesNotExist, RateLimit
 from core.github_service import GithubInformationService
@@ -122,30 +121,3 @@ class LanguageViewSet(mixins.ListModelMixin,
     queryset = Language.objects.all()
     serializer_class = LanguageSerializer
     pagination_class = IdOrderingPagination
-
-
-class TierRankViewSet(mixins.ListModelMixin,
-                      viewsets.GenericViewSet):
-    """
-    endpoint : githubs/tier/
-    """
-
-    queryset = GithubUser.objects.all()
-    serializer_class = TierSerializer
-    pagination_class = TierOrderingPagination
-    filter_backends = [DjangoFilterBackend]
-    filterset_fields = ['tier']
-
-
-class UserRankViewSet(mixins.ListModelMixin,
-                      viewsets.GenericViewSet):
-    """
-    endpoint : githubs/user_rank/
-    """
-
-    # todo: 랭킹 도메인으로 이동
-    queryset = GithubUser.objects.all()
-    serializer_class = TierSerializer
-    pagination_class = UserRankOrderingPagination
-    filter_backends = [DjangoFilterBackend]
-    filterset_fields = ['user_rank']
